add quantity of product

+++ src/main/java/com/wh/service/ShipmentService.java
@@ -9,5 +9,5 @@ public interface ShipmentService {
 	List<Shipment> findAll();
 
 	void save(String date, Long contragentId, Long productId,
-			Double productCount, Long storeId, Long transportId, Long addressId, Boolean paymentType, String comment);
+			Integer productCount, Long storeId, Long transportId, Long addressId, Boolean paymentType, String comment);
 }
