+++ src/main/java/com/wh/web/ShipmentController.java
@@ -58,7 +58,7 @@ public class ShipmentController {
 	
 	@RequestMapping({"/save"})
 	public String addIncoming(HttpSession session, @RequestParam("date") String date, @RequestParam("contragent") Long contragentId,
-			@RequestParam("product") Long productId, @RequestParam("productCount") Double productCount, @RequestParam("store") Long storeId,
+			@RequestParam("product") Long productId, @RequestParam("productCount") Integer productCount, @RequestParam("store") Long storeId,
 			@RequestParam("transport") Long transportId, @RequestParam("address") Long addressId, @RequestParam("comment") String comment,
 			@RequestParam("paymentType") Boolean paymentType) {
 		shipmentService.save(date, contragentId, productId, productCount, storeId, transportId, addressId, paymentType, comment);
